requirements.txt
appstoreconnect==0.10.0
python-dotenv==1.0.0
requests==2.31.0
click==8.1.7
pytz==2024.1
slack-sdk==3.27.1
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.12.0
black==24.3.0
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
tox==4.15.0
pre-commit==3.6.2
types-requests==2.31.0.20240311
types-pytz==2024.1.0.20240203
PyJWT>=1.6.4,<2.0.0
cryptography==42.0.5

# Development dependencies
schedule==1.2.1
python-json-logger==2.0.7
flake8-docstrings==1.7.0
tox==4.15.0 